Split folder browser tests by concern
