+++ src/gui_app/folder_browser.rs
@@ -508,998 +508,4 @@ fn plural(count: usize) -> &'static str {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        FolderBrowserMessage, FolderBrowserState, FolderDragPreview, FolderScanDiscoveryBatch,
-        MIN_FILE_COLUMN_WIDTH, path_id, scan_source_with_progress,
-    };
-    use radiant::{
-        layout::Point,
-        runtime::ExternalDragPayload,
-        widgets::{DragHandleMessage, PointerModifiers, TextInputMessage},
-    };
-    use std::{fs, path::PathBuf};
-
-    #[test]
-    fn source_scan_installs_finished_tree_after_placeholder_selection() {
-        let root = temp_source_root("radiant-gui-source-scan");
-        fs::create_dir_all(root.join("drums")).expect("create nested folder");
-        fs::write(root.join("drums").join("kick.wav"), [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::load_default();
-        let request = browser
-            .begin_add_source_path(root.clone(), 42)
-            .expect("new source should request scan");
-
-        assert_eq!(browser.root_path(), root.as_path());
-        assert!(browser.selected_audio_files().is_empty());
-
-        let mut progress_events = Vec::new();
-        let mut discovery_events = Vec::new();
-        let result = scan_source_with_progress(
-            request,
-            |progress| progress_events.push(progress),
-            |event| discovery_events.push(event),
-        );
-        assert!(browser.apply_scan_finished(result));
-
-        browser.begin_select_source(root.to_string_lossy().to_string(), 43);
-        browser.activate_folder(path_id(&root.join("drums")));
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["kick.wav"]
-        );
-        assert!(
-            progress_events
-                .iter()
-                .any(|progress| progress.phase == "Scanning" && progress.total == 0)
-        );
-        assert!(!discovery_events.is_empty());
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn non_wav_audio_looking_files_are_visible_but_not_supported_audio() {
-        let root = temp_source_root("radiant-gui-unsupported-audio");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        for name in ["kick.wav", "loop.aif", "loop.aiff", "loop.flac", "loop.mp3"] {
-            fs::write(drums.join(name), [0_u8; 8]).expect("write audio-looking file");
-        }
-
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["kick.wav"]
-        );
-
-        let unsupported = browser
-            .selected_files()
-            .iter()
-            .filter(|file| file.kind == "Unsupported audio")
-            .map(|file| file.name.as_str())
-            .collect::<Vec<_>>();
-        assert_eq!(
-            unsupported,
-            vec!["loop.aif", "loop.aiff", "loop.flac", "loop.mp3"]
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn source_scan_discoveries_populate_selected_tree_before_finish() {
-        let root = temp_source_root("radiant-gui-source-streaming");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        fs::write(drums.join("kick.wav"), [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::load_default();
-        let request = browser
-            .begin_add_source_path(root.clone(), 77)
-            .expect("new source should request scan");
-
-        let mut progress_events = Vec::new();
-        let mut discovery_events = Vec::new();
-        let result = scan_source_with_progress(
-            request,
-            |progress| progress_events.push(progress),
-            |event| discovery_events.push(event),
-        );
-
-        for event in discovery_events {
-            browser.apply_scan_discovered(event);
-        }
-        browser.activate_folder(path_id(&drums));
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["kick.wav"]
-        );
-
-        assert!(browser.apply_scan_finished(result));
-        assert!(progress_events.iter().all(|progress| progress.total == 0));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn batched_scan_discoveries_clone_selected_tree_once_per_batch() {
-        let root = temp_source_root("radiant-gui-source-batch");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        fs::write(drums.join("kick.wav"), [0_u8; 8]).expect("write wav");
-        fs::write(drums.join("snare.wav"), [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::load_default();
-        let request = browser
-            .begin_add_source_path(root.clone(), 88)
-            .expect("new source should request scan");
-
-        let mut discovery_events = Vec::new();
-        let result =
-            scan_source_with_progress(request, |_| {}, |event| discovery_events.push(event));
-        assert!(
-            browser.apply_scan_discovered_batch(FolderScanDiscoveryBatch {
-                task_id: 88,
-                source_id: path_id(&root),
-                events: discovery_events,
-            })
-        );
-        browser.activate_folder(path_id(&drums));
-        assert_eq!(browser.selected_audio_files().len(), 2);
-
-        assert!(browser.apply_scan_finished(result));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn visible_folder_depths_are_stable_for_siblings() {
-        let root = temp_source_root("radiant-gui-folder-depths");
-        for child in ["alpha", "beta", "gamma"] {
-            fs::create_dir_all(root.join("parent").join(child)).expect("create nested folder");
-        }
-        let browser = FolderBrowserState::from_root(root.clone());
-        let mut browser = browser;
-        browser.activate_folder(path_id(&root.join("parent")));
-
-        let sibling_depths = browser
-            .visible_folders()
-            .into_iter()
-            .filter(|folder| ["alpha", "beta", "gamma"].contains(&folder.name.as_str()))
-            .map(|folder| folder.depth)
-            .collect::<Vec<_>>();
-
-        assert_eq!(sibling_depths, vec![2, 2, 2]);
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_keyboard_navigation_moves_visible_selection_and_expands_collapses() {
-        let root = temp_source_root("radiant-gui-folder-keyboard");
-        let drums = root.join("drums");
-        let kicks = drums.join("kicks");
-        let snares = drums.join("snares");
-        fs::create_dir_all(&kicks).expect("create kicks folder");
-        fs::create_dir_all(&snares).expect("create snares folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-
-        assert_eq!(browser.selected_folder, path_id(&root));
-        assert!(browser.navigate_selected_folder(1));
-        assert_eq!(browser.selected_folder, path_id(&drums));
-        assert!(!browser.is_expanded(&path_id(&drums)));
-        assert!(browser.expand_selected_folder());
-        assert!(browser.is_expanded(&path_id(&drums)));
-        assert!(browser.collapse_selected_folder());
-        assert!(!browser.is_expanded(&path_id(&drums)));
-        assert!(browser.expand_selected_folder());
-        assert!(browser.is_expanded(&path_id(&drums)));
-        assert!(browser.navigate_selected_folder(1));
-        assert_eq!(browser.selected_folder, path_id(&kicks));
-        assert!(browser.navigate_selected_folder(1));
-        assert_eq!(browser.selected_folder, path_id(&snares));
-        assert!(!browser.navigate_selected_folder(1));
-        assert_eq!(browser.selected_folder, path_id(&snares));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_keyboard_navigation_moves_audio_selection_without_leaving_folder() {
-        let root = temp_source_root("radiant-gui-file-keyboard");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let hat = drums.join("hat.wav");
-        let kick = drums.join("kick.wav");
-        let snare = drums.join("snare.wav");
-        fs::write(&hat, [0_u8; 8]).expect("write hat");
-        fs::write(&kick, [0_u8; 8]).expect("write kick");
-        fs::write(&snare, [0_u8; 8]).expect("write snare");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&hat));
-
-        assert_eq!(browser.navigate_vertical(1, false), Some(path_id(&kick)));
-        browser.select_file(path_id(&kick));
-        assert_eq!(browser.navigate_vertical(1, false), Some(path_id(&snare)));
-        browser.select_file(path_id(&snare));
-        assert_eq!(browser.navigate_vertical(1, false), None);
-        assert_eq!(browser.selected_file_id(), Some(path_id(&snare).as_str()));
-        assert_eq!(browser.navigate_vertical(-1, false), Some(path_id(&kick)));
-        assert_eq!(browser.selected_folder, path_id(&drums));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_keyboard_navigation_can_extend_audio_selection() {
-        let root = temp_source_root("radiant-gui-file-keyboard-extend");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let hat = drums.join("hat.wav");
-        let kick = drums.join("kick.wav");
-        let snare = drums.join("snare.wav");
-        fs::write(&hat, [0_u8; 8]).expect("write hat");
-        fs::write(&kick, [0_u8; 8]).expect("write kick");
-        fs::write(&snare, [0_u8; 8]).expect("write snare");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&hat));
-
-        assert_eq!(browser.navigate_vertical(1, true), Some(path_id(&kick)));
-        assert_eq!(browser.navigate_vertical(1, true), Some(path_id(&snare)));
-
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![hat.clone(), kick.clone(), snare.clone()]
-        );
-        assert_eq!(browser.selected_file_id(), Some(path_id(&snare).as_str()));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_mouse_selection_toggles_and_extends_audio_selection() {
-        let root = temp_source_root("radiant-gui-file-mouse-multi-select");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let hat = drums.join("hat.wav");
-        let kick = drums.join("kick.wav");
-        let snare = drums.join("snare.wav");
-        let tom = drums.join("tom.wav");
-        for file in [&hat, &kick, &snare, &tom] {
-            fs::write(file, [0_u8; 8]).expect("write wav");
-        }
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&hat));
-
-        browser.select_file_with_modifiers(
-            path_id(&snare),
-            PointerModifiers {
-                command: true,
-                ..Default::default()
-            },
-        );
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![hat.clone(), snare.clone()]
-        );
-
-        browser.select_file_with_modifiers(
-            path_id(&tom),
-            PointerModifiers {
-                shift: true,
-                ..Default::default()
-            },
-        );
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![snare.clone(), tom.clone()]
-        );
-
-        browser.select_file_with_modifiers(
-            path_id(&kick),
-            PointerModifiers {
-                command: true,
-                shift: true,
-                ..Default::default()
-            },
-        );
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![kick.clone(), snare.clone(), tom.clone()]
-        );
-
-        browser.select_file_with_modifiers(
-            path_id(&snare),
-            PointerModifiers {
-                command: true,
-                ..Default::default()
-            },
-        );
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![kick.clone(), tom.clone()]
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_keyboard_navigation_follow_window_moves_only_near_edges() {
-        let root = temp_source_root("radiant-gui-file-follow-window");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let files = (0..20)
-            .map(|index| drums.join(format!("sample_{index:02}.wav")))
-            .collect::<Vec<_>>();
-        for file in &files {
-            fs::write(file, [0_u8; 8]).expect("write wav");
-        }
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&files[4]));
-
-        let window = browser.follow_selected_file_view(6, 1, 1);
-        assert_eq!(window.viewport_start, 1);
-        assert_eq!(browser.file_view_start(), 1);
-
-        assert_eq!(
-            browser.navigate_vertical(1, false),
-            Some(path_id(&files[5]))
-        );
-        let window = browser.follow_selected_file_view(6, 1, 1);
-        assert_eq!(window.viewport_start, 2);
-        assert_eq!(browser.file_view_start(), 2);
-
-        assert_eq!(
-            browser.navigate_vertical(1, false),
-            Some(path_id(&files[6]))
-        );
-        let window = browser.follow_selected_file_view(6, 1, 1);
-        assert_eq!(window.viewport_start, 3);
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_scroll_tracking_allows_runtime_clamped_bottom_offsets() {
-        let root = temp_source_root("radiant-gui-file-scroll-bottom");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let files = (0..24)
-            .map(|index| drums.join(format!("sample_{index:02}.wav")))
-            .collect::<Vec<_>>();
-        for file in &files {
-            fs::write(file, [0_u8; 8]).expect("write wav");
-        }
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        browser.set_file_view_start_from_scroll_offset(23.0 * 22.0, 22.0);
-
-        assert_eq!(browser.file_view_start(), 23);
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn select_all_audio_files_selects_current_folder_samples() {
-        let root = temp_source_root("radiant-gui-file-select-all");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let hat = drums.join("hat.wav");
-        let kick = drums.join("kick.wav");
-        let note = drums.join("note.txt");
-        fs::write(&hat, [0_u8; 8]).expect("write hat");
-        fs::write(&kick, [0_u8; 8]).expect("write kick");
-        fs::write(&note, [0_u8; 8]).expect("write note");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        assert_eq!(browser.select_all_audio_files(), 2);
-
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![hat.clone(), kick.clone()]
-        );
-        assert!(!browser.is_file_selected(&path_id(&note)));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn sample_file_sort_toggles_by_column_and_navigation_uses_sorted_order() {
-        let root = temp_source_root("radiant-gui-file-sort");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let small = drums.join("small.wav");
-        let large = drums.join("large.wav");
-        fs::write(&small, [0_u8; 8]).expect("write small");
-        fs::write(&large, [0_u8; 128]).expect("write large");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        browser.apply_message(FolderBrowserMessage::SortFileColumn(String::from("size")));
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["small.wav", "large.wav"]
-        );
-
-        browser.apply_message(FolderBrowserMessage::SortFileColumn(String::from("size")));
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["large.wav", "small.wav"]
-        );
-        browser.select_file(path_id(&large));
-        assert_eq!(browser.navigate_vertical(1, false), Some(path_id(&small)));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn sample_file_column_resize_clamps_width() {
-        let mut browser = FolderBrowserState::load_default();
-
-        browser.apply_message(FolderBrowserMessage::ResizeFileColumn(
-            String::from("extension"),
-            radiant::widgets::DragHandleMessage::Started {
-                position: Point::new(100.0, 0.0),
-            },
-        ));
-        browser.apply_message(FolderBrowserMessage::ResizeFileColumn(
-            String::from("extension"),
-            radiant::widgets::DragHandleMessage::Moved {
-                position: Point::new(-200.0, 0.0),
-            },
-        ));
-
-        let extension_width = browser
-            .visible_file_columns()
-            .into_iter()
-            .find(|column| column.id == "extension")
-            .map(|column| column.width)
-            .unwrap();
-        assert_eq!(extension_width, MIN_FILE_COLUMN_WIDTH);
-    }
-
-    #[test]
-    fn browser_keyboard_navigation_is_disabled_while_renaming() {
-        let root = temp_source_root("radiant-gui-keyboard-rename");
-        let drums = root.join("drums");
-        let kicks = drums.join("kicks");
-        fs::create_dir_all(&kicks).expect("create kicks folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser
-            .begin_rename_selected()
-            .expect("rename can start")
-            .expect("rename input id");
-
-        assert!(browser.rename_active());
-        assert_eq!(browser.navigate_vertical(1, false), None);
-        assert!(!browser.expand_selected_folder());
-        assert!(!browser.collapse_selected_folder());
-        assert_eq!(browser.selected_folder, path_id(&drums));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_rename_updates_filesystem_tree_and_selected_audio_files() {
-        let root = temp_source_root("radiant-gui-folder-rename");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        fs::write(drums.join("kick.wav"), [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        let input_id = browser
-            .begin_rename_selected()
-            .expect("rename can start")
-            .expect("rename input id");
-        assert_ne!(input_id, 0);
-        let status = browser
-            .apply_rename_input(TextInputMessage::Submitted {
-                value: String::from("breaks"),
-            })
-            .expect("rename status");
-
-        assert_eq!(status, "Renamed folder to breaks");
-        assert!(!drums.exists());
-        assert!(root.join("breaks").join("kick.wav").is_file());
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.id.as_str())
-                .collect::<Vec<_>>(),
-            vec![path_id(&root.join("breaks").join("kick.wav"))]
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn create_subfolder_starts_pending_rename_row_and_creates_on_submit() {
-        let root = temp_source_root("radiant-gui-folder-create");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        let input_id = browser
-            .begin_create_subfolder()
-            .expect("create can start")
-            .expect("rename input id");
-        let pending = drums.join("New folder");
-
-        assert_ne!(input_id, 0);
-        assert!(!pending.exists());
-        assert!(browser.is_expanded(&path_id(&drums)));
-        assert!(
-            browser.visible_folders().into_iter().any(|folder| {
-                folder.id == path_id(&pending)
-                    && folder.selected
-                    && folder.rename_draft.as_deref() == Some("New folder")
-                    && folder.rename_input_id == Some(input_id)
-            }),
-            "expected pending child rename row"
-        );
-
-        let status = browser
-            .apply_rename_input(TextInputMessage::Submitted {
-                value: String::from("loops"),
-            })
-            .expect("create status");
-
-        assert_eq!(status, "Created folder loops");
-        assert!(!pending.exists());
-        assert!(drums.join("loops").is_dir());
-        assert_eq!(browser.selected_folder, path_id(&drums.join("loops")));
-        assert!(
-            browser
-                .visible_folders()
-                .into_iter()
-                .any(|folder| folder.id == path_id(&drums.join("loops"))
-                    && folder.name == "loops"
-                    && folder.rename_draft.is_none())
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn create_subfolder_cancel_removes_pending_row_without_touching_disk() {
-        let root = temp_source_root("radiant-gui-folder-create-cancel");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        browser
-            .begin_create_subfolder()
-            .expect("create can start")
-            .expect("rename input id");
-        let pending = drums.join("New folder");
-
-        browser.activate_folder(path_id(&drums));
-
-        assert!(!pending.exists());
-        assert_eq!(browser.selected_folder, path_id(&drums));
-        assert!(
-            browser
-                .visible_folders()
-                .into_iter()
-                .all(|folder| folder.id != path_id(&pending))
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn create_subfolder_default_name_skips_existing_folder() {
-        let root = temp_source_root("radiant-gui-folder-create-unique");
-        let drums = root.join("drums");
-        fs::create_dir_all(drums.join("New folder")).expect("create existing folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-
-        browser
-            .begin_create_subfolder()
-            .expect("create can start")
-            .expect("rename input id");
-
-        assert!(
-            browser.visible_folders().into_iter().any(|folder| {
-                folder.id == path_id(&drums.join("New folder 2"))
-                    && folder.rename_draft.as_deref() == Some("New folder 2")
-            }),
-            "expected unique default name"
-        );
-        assert!(!drums.join("New folder 2").exists());
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_rename_hides_and_preserves_extension() {
-        let root = temp_source_root("radiant-gui-file-rename");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        let kick = drums.join("kick loop.wav");
-        fs::write(&kick, [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&kick));
-
-        let input_id = browser
-            .begin_rename_selected()
-            .expect("rename can start")
-            .expect("rename input id");
-        let rename = browser
-            .file_rename_view(&path_id(&kick))
-            .expect("file rename view");
-        assert_eq!(rename.input_id, input_id);
-        assert_eq!(rename.draft, "kick loop");
-        assert_eq!(rename.selection_start, 0);
-        assert_eq!(rename.selection_end, "kick loop".chars().count());
-
-        let status = browser
-            .apply_rename_input(TextInputMessage::Submitted {
-                value: String::from("snare loop"),
-            })
-            .expect("rename status");
-
-        assert_eq!(status, "Renamed file to snare loop.wav");
-        assert!(!kick.exists());
-        assert!(drums.join("snare loop.wav").is_file());
-        assert_eq!(
-            browser.selected_file_id(),
-            Some(path_id(&drums.join("snare loop.wav")).as_str())
-        );
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| (
-                    file.name.as_str(),
-                    file.stem.as_str(),
-                    file.extension.as_str()
-                ))
-                .collect::<Vec<_>>(),
-            vec![("snare loop.wav", "snare loop", "wav")]
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_rename_submission_cannot_change_extension() {
-        let root = temp_source_root("radiant-gui-file-rename-extension");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create nested folder");
-        let kick = drums.join("kick.wav");
-        fs::write(&kick, [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&kick));
-        browser
-            .begin_rename_selected()
-            .expect("rename can start")
-            .expect("rename input id");
-
-        let status = browser
-            .apply_rename_input(TextInputMessage::Submitted {
-                value: String::from("snare.aiff"),
-            })
-            .expect("rename status");
-
-        assert_eq!(status, "Renamed file to snare.aiff.wav");
-        assert!(!kick.exists());
-        assert!(drums.join("snare.aiff.wav").is_file());
-        assert!(!drums.join("snare.aiff").exists());
-        assert_eq!(
-            browser.selected_file_id(),
-            Some(path_id(&drums.join("snare.aiff.wav")).as_str())
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn root_folder_rename_is_rejected_from_tree() {
-        let root = temp_source_root("radiant-gui-root-rename");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-
-        assert_eq!(
-            browser.begin_rename_selected(),
-            Err(String::from("Select a subfolder to rename"))
-        );
-        assert!(root.is_dir());
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_delete_blocks_hard_delete_and_keeps_selected_folder() {
-        let root = temp_source_root("radiant-gui-folder-delete");
-        let drums = root.join("drums");
-        let kicks = drums.join("kicks");
-        fs::create_dir_all(&kicks).expect("create nested folder");
-        fs::write(kicks.join("kick.wav"), [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.expand_selected_folder();
-        browser.activate_folder(path_id(&kicks));
-
-        let target = browser
-            .selected_delete_target()
-            .expect("subfolder can be deleted");
-        assert_eq!(target.name, "kicks");
-        let error = browser
-            .delete_selected_folder()
-            .expect_err("hard delete should be blocked");
-
-        assert_eq!(
-            error,
-            "Trash workflow is not available in the default GUI yet; no folder was deleted"
-        );
-        assert!(kicks.exists());
-        assert_eq!(browser.selected_folder, path_id(&kicks));
-        assert!(browser.find_folder(&path_id(&kicks)).is_some());
-        assert!(browser.expanded_folders.contains(&path_id(&drums)));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_delete_blocks_hard_delete_and_keeps_selection() {
-        let root = temp_source_root("radiant-gui-file-delete");
-        let drums = root.join("drums");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        let hat = drums.join("hat.wav");
-        let kick = drums.join("kick.wav");
-        let snare = drums.join("snare.wav");
-        for file in [&hat, &kick, &snare] {
-            fs::write(file, [0_u8; 8]).expect("write wav");
-        }
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&kick));
-
-        let target = browser
-            .selected_file_delete_target()
-            .expect("selected file can be deleted");
-        assert_eq!(target.names, vec![String::from("kick.wav")]);
-        let error = browser
-            .delete_selected_files()
-            .expect_err("hard delete should be blocked");
-
-        assert_eq!(
-            error,
-            "Trash workflow is not available in the default GUI yet; no files were deleted"
-        );
-        assert!(kick.exists());
-        assert!(
-            browser
-                .selected_files()
-                .iter()
-                .any(|file| file.name == "kick.wav")
-        );
-        assert_eq!(browser.selected_file_id(), Some(path_id(&kick).as_str()));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_drag_drop_moves_subtree_into_target_folder() {
-        let root = temp_source_root("radiant-gui-folder-drag-drop");
-        let kicks = root.join("drums").join("kicks");
-        let loops = root.join("loops");
-        fs::create_dir_all(&kicks).expect("create kicks folder");
-        fs::create_dir_all(&loops).expect("create loops folder");
-        let kick = kicks.join("kick.wav");
-        fs::write(&kick, [0_u8; 8]).expect("write wav");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&root.join("drums")));
-        browser.expand_selected_folder();
-        browser.activate_folder(path_id(&kicks));
-        browser.select_file(path_id(&kick));
-
-        browser.apply_folder_drag(
-            path_id(&kicks),
-            DragHandleMessage::Started {
-                position: Point::new(0.0, 0.0),
-            },
-        );
-        let result = browser
-            .drop_drag_on_folder(&path_id(&loops))
-            .expect("folder drag/drop should move");
-
-        let moved_kicks = loops.join("kicks");
-        let moved_kick = moved_kicks.join("kick.wav");
-        assert_eq!(
-            result.moved_paths,
-            vec![(kicks.clone(), moved_kicks.clone())]
-        );
-        assert!(!kicks.exists());
-        assert!(moved_kick.is_file());
-        assert_eq!(browser.selected_folder, path_id(&moved_kicks));
-        assert_eq!(
-            browser.selected_file_id(),
-            Some(path_id(&moved_kick).as_str())
-        );
-        assert!(browser.find_folder(&path_id(&moved_kicks)).is_some());
-        assert!(browser.expanded_folders.contains(&path_id(&loops)));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_drag_preview_tracks_pointer_and_hover_target() {
-        let root = temp_source_root("radiant-gui-folder-drag-preview");
-        let kicks = root.join("drums").join("kicks");
-        let loops = root.join("loops");
-        fs::create_dir_all(&kicks).expect("create kicks folder");
-        fs::create_dir_all(&loops).expect("create loops folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&root.join("drums")));
-        browser.expand_selected_folder();
-
-        browser.apply_folder_drag(
-            path_id(&kicks),
-            DragHandleMessage::Started {
-                position: Point::new(10.0, 20.0),
-            },
-        );
-        assert_eq!(
-            browser.drag_preview(),
-            Some(FolderDragPreview {
-                label: String::from("kicks"),
-                pointer: Point::new(10.0, 20.0),
-            })
-        );
-
-        browser.apply_folder_drag(
-            path_id(&kicks),
-            DragHandleMessage::Moved {
-                position: Point::new(30.0, 42.0),
-            },
-        );
-        assert_eq!(
-            browser.drag_preview().map(|preview| preview.pointer),
-            Some(Point::new(30.0, 42.0))
-        );
-
-        browser.apply_message(FolderBrowserMessage::HoverDropTarget(path_id(&loops)));
-        let hovered = browser
-            .visible_folders()
-            .into_iter()
-            .find(|folder| folder.id == path_id(&loops))
-            .expect("loops folder visible");
-        assert!(hovered.drop_candidate);
-        assert!(hovered.drop_target);
-
-        browser.apply_message(FolderBrowserMessage::HoverDropTarget(path_id(&kicks)));
-        assert!(
-            browser
-                .visible_folders()
-                .into_iter()
-                .all(|folder| !folder.drop_target)
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn folder_drag_external_request_uses_preview_label_and_paths() {
-        let root = temp_source_root("radiant-gui-folder-external-drag");
-        let kicks = root.join("drums").join("kicks");
-        fs::create_dir_all(&kicks).expect("create kicks folder");
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&root.join("drums")));
-        browser.expand_selected_folder();
-
-        browser.apply_folder_drag(
-            path_id(&kicks),
-            DragHandleMessage::Started {
-                position: Point::new(10.0, 20.0),
-            },
-        );
-        let request = browser
-            .external_drag_request()
-            .expect("folder drag should expose external request");
-
-        assert_eq!(request.preview.label, "kicks");
-        assert_eq!(
-            request.payload,
-            ExternalDragPayload::Files(vec![kicks.clone()])
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn file_drag_drop_moves_selected_files_into_target_folder() {
-        let root = temp_source_root("radiant-gui-file-drag-drop");
-        let drums = root.join("drums");
-        let loops = root.join("loops");
-        fs::create_dir_all(&drums).expect("create drums folder");
-        fs::create_dir_all(&loops).expect("create loops folder");
-        let kick = drums.join("kick.wav");
-        let snare = drums.join("snare.wav");
-        let hat = drums.join("hat.wav");
-        for file in [&kick, &snare, &hat] {
-            fs::write(file, [0_u8; 8]).expect("write wav");
-        }
-        let mut browser = FolderBrowserState::from_root(root.clone());
-        browser.activate_folder(path_id(&drums));
-        browser.select_file(path_id(&kick));
-        browser.select_file_with_modifiers(
-            path_id(&snare),
-            PointerModifiers {
-                command: true,
-                ..Default::default()
-            },
-        );
-
-        browser.begin_file_drag(path_id(&kick), Point::new(4.0, 8.0));
-        let result = browser
-            .drop_drag_on_folder(&path_id(&loops))
-            .expect("file drag/drop should move");
-
-        let moved_kick = loops.join("kick.wav");
-        let moved_snare = loops.join("snare.wav");
-        assert_eq!(result.moved_paths.len(), 2);
-        assert!(!kick.exists());
-        assert!(!snare.exists());
-        assert!(hat.is_file());
-        assert!(moved_kick.is_file());
-        assert!(moved_snare.is_file());
-        assert_eq!(browser.selected_folder, path_id(&loops));
-        assert_eq!(
-            browser.selected_file_paths(),
-            vec![moved_kick.clone(), moved_snare.clone()]
-        );
-        assert_eq!(
-            browser
-                .selected_audio_files()
-                .iter()
-                .map(|file| file.name.as_str())
-                .collect::<Vec<_>>(),
-            vec!["kick.wav", "snare.wav"]
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn root_folder_delete_is_rejected_from_tree() {
-        let root = temp_source_root("radiant-gui-root-delete");
-        let browser = FolderBrowserState::from_root(root.clone());
-
-        assert_eq!(
-            browser.selected_delete_target(),
-            Err(String::from("Root folder cannot be deleted"))
-        );
-        assert!(root.is_dir());
-        let _ = fs::remove_dir_all(root);
-    }
-
-    fn temp_source_root(name: &str) -> PathBuf {
-        let root = std::env::temp_dir().join(format!(
-            "{name}-{}",
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .expect("clock")
-                .as_nanos()
-        ));
-        fs::create_dir_all(&root).expect("create temp root");
-        root
-    }
-}
+mod tests;
